+++ pipelines/pipelines_nuno/__init__.py
@@ -3,4 +3,4 @@
 Flows para o exercicio de esquenta (SME)
 """
 
-from pipelines.pipelines_nuno.aula_4.flow import *
+from pipelines.pipelines_nuno.aula_4.flows import *
